+++ apps/web/src/config/nodes.ts
@@ -20,6 +20,8 @@ const POLYGON_ZKEVM_NODES = [
 
 const ARBITRUM_NODES = [
   ...arbitrum.rpcUrls.public.http,
+  'https://arbitrum-one.publicnode.com',
+  'https://arbitrum.llamarpc.com',
   process.env.NEXT_PUBLIC_QUICK_NODE_ARB_PRODUCTION || '',
   getNodeRealUrlV2(ChainId.ARBITRUM_ONE, process.env.NEXT_PUBLIC_NODE_REAL_API_ETH) || '',
 ]
@@ -27,12 +29,15 @@ const ARBITRUM_NODES = [
 export const SERVER_NODES = {
   [ChainId.BSC]: [
     process.env.NEXT_PUBLIC_NODE_PRODUCTION || '',
+    'https://bsc.publicnode.com',
+    'https://binance.llamarpc.com',
     'https://bsc-dataseed1.defibit.io',
     'https://bsc-dataseed1.binance.org',
   ].filter(Boolean),
   [ChainId.BSC_TESTNET]: ['https://data-seed-prebsc-1-s1.binance.org:8545'],
   [ChainId.ETHEREUM]: [
     getNodeRealUrlV2(ChainId.ETHEREUM, process.env.SERVER_NODE_REAL_API_ETH) || '',
+    'https://ethereum.publicnode.com',
     'https://eth.llamarpc.com',
     'https://cloudflare-eth.com',
   ],
@@ -47,7 +52,10 @@ export const SERVER_NODES = {
     'https://polygon-zkevm-testnet.rpc.thirdweb.com',
     ...polygonZkEvmTestnet.rpcUrls.public.http,
   ],
-  [ChainId.ZKSYNC]: zkSync.rpcUrls.public.http,
+  [ChainId.ZKSYNC]: [
+    ...zkSync.rpcUrls.public.http,
+    getNodeRealUrlV2(ChainId.ZKSYNC, process.env.SERVER_NODE_REAL_API_ETH) || '',
+  ],
   [ChainId.ZKSYNC_TESTNET]: zkSyncTestnet.rpcUrls.public.http,
   [ChainId.LINEA]: linea.rpcUrls.public.http,
   [ChainId.LINEA_TESTNET]: [
@@ -56,7 +64,10 @@ export const SERVER_NODES = {
     'https://consensys-zkevm-goerli-prealpha.infura.io/v3/93e8a17747e34ec0ac9a554c1b403965',
   ],
   [ChainId.OPBNB_TESTNET]: opbnbTestnet.rpcUrls.public.http,
-  [ChainId.OPBNB]: opbnb.rpcUrls.public.http,
+  [ChainId.OPBNB]: [
+    ...opbnb.rpcUrls.public.http,
+    getNodeRealUrlV2(ChainId.OPBNB, process.env.SERVER_NODE_REAL_API_ETH) || '',
+  ],
   [ChainId.BASE]: [
     'https://base.publicnode.com',
     // process.env.NEXT_PUBLIC_NODE_REAL_BASE_PRODUCTION,
@@ -70,12 +81,15 @@ export const PUBLIC_NODES = {
   [ChainId.BSC]: [
     process.env.NEXT_PUBLIC_NODE_PRODUCTION || '',
     getNodeRealUrlV2(ChainId.BSC, process.env.NEXT_PUBLIC_NODE_REAL_API_ETH) || '',
+    'https://bsc.publicnode.com',
+    'https://binance.llamarpc.com',
     'https://bsc-dataseed1.defibit.io',
     'https://bsc-dataseed1.binance.org',
   ].filter(Boolean),
   [ChainId.BSC_TESTNET]: ['https://data-seed-prebsc-1-s1.binance.org:8545'],
   [ChainId.ETHEREUM]: [
     getNodeRealUrlV2(ChainId.ETHEREUM, process.env.NEXT_PUBLIC_NODE_REAL_API_ETH) || '',
+    'https://ethereum.publicnode.com',
     'https://eth.llamarpc.com',
     'https://cloudflare-eth.com',
   ].filter(Boolean),
@@ -93,7 +107,10 @@ export const PUBLIC_NODES = {
     ...polygonZkEvmTestnet.rpcUrls.public.http,
     'https://polygon-zkevm-testnet.rpc.thirdweb.com',
   ],
-  [ChainId.ZKSYNC]: zkSync.rpcUrls.public.http,
+  [ChainId.ZKSYNC]: [
+    ...zkSync.rpcUrls.public.http,
+    getNodeRealUrlV2(ChainId.ZKSYNC, process.env.NEXT_PUBLIC_NODE_REAL_API_ETH) || '',
+  ],
   [ChainId.ZKSYNC_TESTNET]: zkSyncTestnet.rpcUrls.public.http,
   [ChainId.LINEA]: linea.rpcUrls.public.http,
   [ChainId.LINEA_TESTNET]: [
@@ -102,7 +119,11 @@ export const PUBLIC_NODES = {
     'https://consensys-zkevm-goerli-prealpha.infura.io/v3/93e8a17747e34ec0ac9a554c1b403965',
   ],
   [ChainId.OPBNB_TESTNET]: opbnbTestnet.rpcUrls.public.http,
-  [ChainId.OPBNB]: opbnb.rpcUrls.public.http,
+  [ChainId.OPBNB]: [
+    ...opbnb.rpcUrls.public.http,
+    getNodeRealUrlV2(ChainId.OPBNB, process.env.NEXT_PUBLIC_NODE_REAL_API_ETH) || '',
+    'https://opbnb.publicnode.com',
+  ],
   [ChainId.BASE]: [
     'https://base.publicnode.com',
     // process.env.NEXT_PUBLIC_NODE_REAL_BASE_PRODUCTION,