+++ apps/web/src/pages/nfts/index.tsx
@@ -1,13 +1,7 @@
 import NftMarket from 'views/Nft/market/Home'
-import SubMenu from 'views/Nft/market/components/SubMenu'
 
 const NftMarketPage = () => {
-  return (
-    <>
-      <SubMenu />
-      <NftMarket />
-    </>
-  )
+  return <NftMarket />
 }
 
 export default NftMarketPage