+++ apps/web/src/pages/swap.tsx
@@ -2,8 +2,8 @@ import { CHAIN_IDS } from 'utils/wagmi'
 
 import { V3SubgraphHealthIndicator } from 'components/SubgraphHealthIndicator'
 
+import { SwapFeaturesProvider } from 'views/Swap/SwapFeaturesContext'
 import Swap from '../views/Swap'
-import { SwapFeaturesProvider } from '../views/Swap/SwapFeaturesContext'
 
 const SwapPage = () => {
   return (