+++ apps/web/src/views/Home/index.tsx
@@ -1,6 +1,5 @@
 import { useTranslation } from '@pancakeswap/localization'
 import { Box, Flex, PageSection, Text, useMatchBreakpoints } from '@pancakeswap/uikit'
-import { useAnniversaryEffect } from 'hooks/useAnniversaryEffect'
 import useTheme from 'hooks/useTheme'
 import { styled } from 'styled-components'
 import MultipleBanner from './components/Banners/MultipleBanner'
@@ -34,7 +33,7 @@ const Home: React.FC<React.PropsWithChildren> = () => {
   const HomeSectionContainerStyles = { margin: '0', width: '100%', maxWidth: '968px', padding: '0px 16px' }
   const { t } = useTranslation()
   const { isMobile } = useMatchBreakpoints()
-  useAnniversaryEffect()
+
   return (
     <Box style={{ width: isMobile ? '100vw' : 'calc(100vw - 8px)', overflow: 'hidden', boxSizing: 'border-box' }}>
       <style jsx global>