+++ packages/farms/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pancakeswap/farms
 
+## 1.0.20
+
+### Patch Changes
+
+- ed3146c93: Add chain names
+- Updated dependencies [ed3146c93]
+  - @pancakeswap/chains@0.2.0
+  - @pancakeswap/sdk@5.7.1
+  - @pancakeswap/tokens@0.5.2
+  - @pancakeswap/v3-sdk@3.5.2
+
 ## 1.0.19
 
 ### Patch Changes